+++ src/app/OIDCClients/data/index.js
@@ -5,11 +5,16 @@ const mapEntity = async (entity) => {
   return {
     friendlyName: entity.name,
     client_id: entity.relyingParty.client_id,
+    token_endpoint_auth_method: entity.relyingParty.token_endpoint_auth_method,
     client_secret: entity.relyingParty.client_secret,
     api_secret: entity.relyingParty.api_secret || undefined,
+    response_types: entity.relyingParty.response_types,
     redirect_uris: entity.relyingParty.redirect_uris,
+    grant_types: entity.relyingParty.grant_types,
+    service_home: entity.relyingParty.service_home,
     post_logout_redirect_uris: entity.relyingParty.post_logout_redirect_uris,
     params: entity.relyingParty.params,
+    postResetUrl: entity.relyingParty.postResetUrl,
   };
 };
