+++ src/infrastructure/applications/index.js
