+++ src/infrastructure/applications/static.js
